requirements.txt
argparse
black
flake8
mypy
numpy>=1.20.0  # To ensure compatibility with OpenCV on Raspberry Pi.
opencv-python~=4.5.3.56
protobuf>=3.18.0,<4
pyproject-flake8
RPi.GPIO
tflite-support==0.4.3
